add $wgPiwikJSFileURL, $wgPiwikPHPFileURL
Setting these globals in a configuration removes the constraint of having to
host piwik.js on the same domain as piwik.php. This is needed for those who
want to host piwik.js on a CDN.

Both variables take precedence over $wgPiwikURL, but $wgPiwikURL is used
if only one is defined. For instance, in the CDN use-case, you only have to
define $wgPiwikJSFileURL.