Writing more readable (hopefully) code for a CDN-hosted piwik.js
This is a rewrite of #14; the JS generated in the PHP part is reduced.

The parameter $wgPiwikPHPFileURL is removed given it was not an explicit
request and it adds complexity in the code; let’s wait someone really
need such a very specific configuration. Moreover it is almost redundant
with $wgPiwikURL, the only added value was you can change the name of the
endpoint (with something else than piwik.php), which is a very specific case.

Added multiple calls to Xml::encodeJsVar to avoid any security issue
even from sysadmin inputs.