Retrieve search results counts
The searched terms are recognized by default, the search category can be configured,
but the search results counts need explicit code. This is achieved here with
two hooks saving the data (term, profile/category, count) and they are then inserted
in the script and image.

There is only a little overhead in non-search pages (most of the pages).