Bump to 3.1.0 for the new feature 'search-count'
This feature automatically report to Piwik the searched terms in the internal
search engine as well as the search category (in MediaWiki the categories
"Content", "Multimedia", "All", "Custom") and the number of results.

Previously it was possible to obtain the two first informations by configuring
Piwik in a specific way (and it worked with the default settings); hence the
real added value here is the third information, and it will work for sure even
if the default settings regarding the searches are not correctly configured.