Rename extension to Matomo (part 2/2) - version 4.0
This second commit renames the various parameters. To keep partial backward
compatibility, the old parameters $wgPiwik… are still valid but the new ones
$wgMatomo… have precedence when they exist.

By the way, the activation of the extension is no more automatic when installed
with Composer; it must be separately activated with `wfLoadExtension( 'Matomo' );`
in LocalSettings.php. This improves compatibility with farms: this extension can
be activated depending on other code instead of being mandatorily activated by
Composer.

Fixes: #21